SQL DB PLAY

+++ src/main/java/com/pl/carrentalservice/branches/BranchRepository.java
@@ -2,5 +2,5 @@ package com.pl.carrentalservice.branches;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface BranchRepository extends JpaRepository<Branch,Long> {
+public interface BranchRepository extends JpaRepository<Branch,Integer> {
 }
